Data loading , Cleaning , Writing - Uber Fares Data

+++ DataIngestion/DataLoader_UberFares.py
@@ -1,5 +1,6 @@
 from config.spark_config import create_spark_session
-from pyspark.sql.types import StructType, StructField, StringType, TimestampType , DoubleType
+from pyspark.sql.types import StructType, StructField, StringType, TimestampType, DoubleType, IntegerType
+
 
 def load_uberFares_data():
     spark = create_spark_session()
@@ -12,7 +13,7 @@ def load_uberFares_data():
     StructField("pickup_latitude", DoubleType(), False),
     StructField("dropoff_longitude", DoubleType(), True),
     StructField("dropoff_latitude", DoubleType(), True),
-    StructField("passenger_count", DoubleType(), True)
+    StructField("passenger_count", IntegerType(), True)
     ])
     filepath = 'C:/Users/HP/uber_project/Data/UberFaresData/uber.csv'
     UberData = spark.read.csv(filepath,schema=schema,header=True)
